Use README filename for default directory markdown
If a directoy contains a file named `README` GitHub will render the
contents by default when the directory is viewed. This is beneficial as
the 'index' or default page of the directory might not be immediately obvious.

I can't find any official documenation about GitHub doing this for all
directories but at the time of writing this is the case.
https://help.github.com/en/github/creating-cloning-and-archiving-repositories/about-readmes